+++ src/main/kotlin/com/mairwunnx/projectessentials/permissions/EntryPoint.kt
@@ -2,7 +2,6 @@ package com.mairwunnx.projectessentials.permissions
 
 import com.mairwunnx.projectessentials.core.EssBase
 import com.mairwunnx.projectessentials.core.configuration.localization.LocalizationConfigurationUtils
-import com.mairwunnx.projectessentials.core.extensions.sendMsg
 import com.mairwunnx.projectessentials.core.localization.processLocalizations
 import com.mairwunnx.projectessentials.permissions.commands.PermissionsCommand
 import com.mairwunnx.projectessentials.permissions.permissions.PermissionBase
@@ -67,11 +66,7 @@ internal class EntryPoint : EssBase() {
                 ) && !PermissionsAPI.hasPermission(
                     player.name.string, "native.event.block.place"
                 ) -> {
-                    sendMsg(
-                        "permissions",
-                        player.commandSource,
-                        "perm.block_break.place"
-                    )
+                    sendMessage(player.commandSource, "block_break.place")
                     event.isCanceled = true
                     return
                 }
@@ -87,11 +82,7 @@ internal class EntryPoint : EssBase() {
             ) && !PermissionsAPI.hasPermission(
                 event.player.name.string, "native.event.block.break"
             ) -> {
-                sendMsg(
-                    "permissions",
-                    event.player.commandSource,
-                    "perm.block_break.restricted"
-                )
+                sendMessage(event.player.commandSource, "block_break.restricted")
                 event.isCanceled = true
                 return
             }