+++ changelog.md
@@ -6,7 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.14.4-1.2.1] - 2020-03-12
+## [1.14.4-1.2.1] - 2020-03-13
+
+### Added
+- Server-side only safe localization.
+- Default user `#server` with all permissions.
+
+### Changed
+- Improved internal code.
+- Dependencies: Kotlin, KotlinxSerialization updated.
+- Dokka and Forge API updated.
 
 ## [1.14.4-1.2.0] - 2020-02-08